@@ -4,7 +4,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.2+2
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.2.6 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
@@ -16,10 +16,9 @@ dependencies:
   hive: ^2.0.5
   path_provider: ^2.0.8
   auto_size_text: ^3.0.0
-  fl_chart: ^0.45.0
+  fl_chart: ^0.69.0
   url_launcher: ^6.0.20
   dio: ^4.0.4
-  package_info_plus: ^1.4.0
 
 dev_dependencies:
   flutter_test:
