@@ -115,7 +115,7 @@ class CurrencyListItem extends StatelessWidget {
       ),
       child: Text(
         title,
-        style: context.textTheme.bodyText1!.copyWith(
+        style: context.textTheme.bodyLarge!.copyWith(
           color: selected ? Colors.white : AppColors.kDark,
         ),
       ),
